stuttgart-etl/pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stuttgart-etl"
version = "0.1.0"
description = "ETL pipeline for Stuttgart geodata (OSM + public datasets)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12",
  "rich>=13",
  "pyyaml>=6.0",
  "requests>=2.31",
  "geopandas>=0.14",
  "shapely>=2.0",
  "pyproj>=3.6",
  "pyogrio>=0.7",
  "pyrosm>=0.6",
  "duckdb>=1.0",
  "pyarrow>=15",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]


